Commit Base Files
CONTRIBUTING.md & README.md were based on my existing tafthorne
/cppcheck-gcc project.  Updated to reflect that this image is intended
for use as part of a Yocto Project build.

Give example of how to use it in the README.md.  Once someone has cloned
a poky work area, they should be able to start a container from this
image and perform a bitbake build.

Dockerfile.bitbake-yocto-debian based on the current essential package
list for 2.4.2 (Rocko) as listed in the reference manual.  This omits
the graphics support and Eclipse integration packages.

I plan to use git submodules to setup work areas instead of (Google's /
Gerrit's) repo tool, so that is also not installed in this image.