Add Link To BitBake Wikipedia Article
The project link for cppcheck was left over from the original copy-paste
of this document.  Replace it with a link to the Wikipedia article on
BitBake.  As BitBake is part of the OpenEmbedded / Yocto build system
there is little other direct explanatory text of what it is on a
dedicate page.